Fix return type mismatch for battery sensor

+++ lib/Sensing/ArduinoBatterySensor.cpp
@@ -3,10 +3,10 @@
 #include "ArduinoBatterySensor.h"
 #include <Arduino.h>
 
-unsigned int ArduinoBatterySensor::get_battery_voltage_adc() {
+uint16_t ArduinoBatterySensor::get_battery_voltage_adc() {
   return analogRead(SNS_BATTERY_VLTG);
 }
 
-unsigned int ArduinoBatterySensor::get_battery_voltage_mVDC() {
+uint16_t ArduinoBatterySensor::get_battery_voltage_mVDC() {
   return analogRead(SNS_BATTERY_VLTG) * ADC_REFERENCE_mV / ADC_MAX;
 }
